backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
websockets==12.0
python-dotenv==1.0.1
sqlmodel==0.0.14
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart>=0.0.7
